fix: 支持 macOS/arm64 构建
三处平台适配，Linux/Windows 行为不变：

- CMakeLists：-march=x86-64 与 fpng 的 -msse4.1 -mpclmul 加 x86 架构判断。
  clang 在 arm64 上遇到这些选项直接报错，无法配置。
- CoverCache：libc++ 的 file_time_type::rep 是 __int128，std::to_string
  无对应重载导致调用二义，显式转 long long。
- FfmpegSession：测试钩子的弱符号在 Mach-O 下需 weak_import 声明，且链接
  期要 -U 显式允许未定义，否则 ld64 报 undefined symbol（ELF 的 weak 语义
  在 Mach-O 上不等价）。链接选项以 INTERFACE 传给消费方。

+++ src/cover/CoverCache.cpp
@@ -153,7 +153,8 @@ struct FileDescriptor
 std::filesystem::path MakeSiblingTempPath(const std::filesystem::path &finalPath)
 {
     static std::atomic_uint64_t tempCounter{0};
-    const auto now = std::filesystem::file_time_type::clock::now().time_since_epoch().count();
+    const auto now = static_cast<long long>(
+        std::filesystem::file_time_type::clock::now().time_since_epoch().count());
     const auto seq = tempCounter.fetch_add(1, std::memory_order_relaxed);
     const std::string name = Utf8Text(finalPath.filename()) + ".tmp." + std::to_string(::getpid()) + "." + std::to_string(now) + "." + std::to_string(seq);
     return finalPath.parent_path() / name;
